Use VirtualTable as a base class for RangeTable
