+++ h2/src/main/org/h2/index/VirtualTableIndex.java
@@ -5,28 +5,19 @@
  */
 package org.h2.index;
 
-import org.h2.command.dml.AllColumnsForPlan;
 import org.h2.engine.Session;
 import org.h2.message.DbException;
 import org.h2.result.Row;
-import org.h2.result.SearchRow;
-import org.h2.result.SortOrder;
-import org.h2.table.FunctionTable;
 import org.h2.table.IndexColumn;
-import org.h2.table.TableFilter;
 import org.h2.table.VirtualTable;
 
 /**
- * An index for a virtual table that returns a result set. Search in this index
- * performs scan over all rows and should be avoided.
+ * An base class for indexes of virtual tables.
  */
-public class VirtualTableIndex extends BaseIndex {
+public abstract class VirtualTableIndex extends BaseIndex {
 
-    private final VirtualTable table;
-
-    public VirtualTableIndex(VirtualTable table, IndexColumn[] columns) {
-        super(table, 0, null, columns, IndexType.createNonUnique(true));
-        this.table = table;
+    protected VirtualTableIndex(VirtualTable table, String name, IndexColumn[] columns) {
+        super(table, 0, name, columns, IndexType.createNonUnique(true));
     }
 
     @Override
@@ -42,31 +33,6 @@ public class VirtualTableIndex extends BaseIndex {
     @Override
     public void remove(Session session, Row row) {
         throw DbException.getUnsupportedException("Virtual table");
-    }
-
-    @Override
-    public boolean isFindUsingFullTableScan() {
-        return true;
-    }
-
-    @Override
-    public Cursor find(Session session, SearchRow first, SearchRow last) {
-        return new VirtualTableCursor(this, first, last, session, table.getResult(session));
-    }
-
-    @Override
-    public double getCost(Session session, int[] masks, TableFilter[] filters, int filter, SortOrder sortOrder,
-            AllColumnsForPlan allColumnsSet) {
-        if (masks != null) {
-            throw DbException.getUnsupportedException("Virtual table");
-        }
-        long expectedRows;
-        if (table.canGetRowCount()) {
-            expectedRows = table.getRowCountApproximation();
-        } else {
-            expectedRows = database.getSettings().estimatedFunctionTableRows;
-        }
-        return expectedRows * 10;
     }
 
     @Override
@@ -90,11 +56,6 @@ public class VirtualTableIndex extends BaseIndex {
     }
 
     @Override
-    public boolean canGetFirstOrLast() {
-        return false;
-    }
-
-    @Override
     public Cursor findFirstOrLast(Session session, boolean first) {
         throw DbException.getUnsupportedException("Virtual table");
     }
@@ -112,16 +73,6 @@ public class VirtualTableIndex extends BaseIndex {
     @Override
     public long getDiskSpaceUsed() {
         return 0;
-    }
-
-    @Override
-    public String getPlanSQL() {
-        return table instanceof FunctionTable ? "function" : "table scan";
-    }
-
-    @Override
-    public boolean canScan() {
-        return false;
     }
 
 }
