+++ h2/src/test/org/h2/test/scripts/datatypes/boolean.sql
@@ -24,7 +24,7 @@ SELECT TYPE_NAME FROM INFORMATION_SCHEMA.COLUMNS WHERE TABLE_NAME = 'TEST';
 >> BOOLEAN
 
 EXPLAIN SELECT CAST(NULL AS BOOLEAN);
->> SELECT UNKNOWN FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT UNKNOWN FROM SYSTEM_RANGE(1, 1) /* range index */
 
 SELECT NOT TRUE A, NOT FALSE B, NOT NULL C, NOT UNKNOWN D;
 > A     B    C    D
