+++ h2/src/test/org/h2/test/scripts/dml/mergeUsing.sql
@@ -29,7 +29,7 @@ EXPLAIN PLAN
             UPDATE SET P.NAME = S.NAME WHERE 2 = 2 WHEN NOT
         MATCHED THEN
             INSERT (ID, NAME) VALUES (S.ID, S.NAME);
->> MERGE INTO "PUBLIC"."PARENT" USING SELECT "X" AS "ID", ('Coco' || "X") AS "NAME" FROM SYSTEM_RANGE(1, 2) /* PUBLIC.RANGE_INDEX */
+>> MERGE INTO "PUBLIC"."PARENT" USING SELECT "X" AS "ID", ('Coco' || "X") AS "NAME" FROM SYSTEM_RANGE(1, 2) /* range index */
 
 DROP TABLE PARENT;
 > ok
