+++ h2/src/test/org/h2/test/scripts/functions/system/cast.sql
@@ -121,4 +121,4 @@ SELECT * FROM (SELECT CAST('2000-01-01 11:11:11.123456789Z' AS TIMESTAMP(9) WITH
 >> 2000-01-01 11:11:11.123456789+00
 
 EXPLAIN SELECT CAST('A' AS VARCHAR(10)), CAST(NULL AS BOOLEAN), CAST(NULL AS VARCHAR), CAST(1 AS INT);
->> SELECT CAST('A' AS VARCHAR(10)), UNKNOWN, CAST(NULL AS VARCHAR), 1 FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT CAST('A' AS VARCHAR(10)), UNKNOWN, CAST(NULL AS VARCHAR), 1 FROM SYSTEM_RANGE(1, 1) /* range index */
