+++ h2/src/test/org/h2/test/scripts/functions/timeanddate/current-time.sql
@@ -25,4 +25,4 @@ SELECT CAST(CURRENT_TIME(9) AS TIME(9)) = LOCALTIME(9);
 >> TRUE
 
 EXPLAIN SELECT CURRENT_TIME, LOCALTIME, CURRENT_TIME(9), LOCALTIME(9);
->> SELECT CURRENT_TIME, LOCALTIME, CURRENT_TIME(9), LOCALTIME(9) FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT CURRENT_TIME, LOCALTIME, CURRENT_TIME(9), LOCALTIME(9) FROM SYSTEM_RANGE(1, 1) /* range index */
