+++ h2/src/test/org/h2/test/scripts/predicates/in.sql
@@ -288,4 +288,4 @@ SELECT * FROM SYSTEM_RANGE(1, 10) WHERE X IN ((SELECT 1), (SELECT 2));
 > rows: 2
 
 EXPLAIN SELECT * FROM SYSTEM_RANGE(1, 10) WHERE X IN ((SELECT 1), (SELECT 2));
->> SELECT "PUBLIC"."SYSTEM_RANGE"."X" FROM SYSTEM_RANGE(1, 10) /* PUBLIC.RANGE_INDEX: X IN((SELECT 1 FROM SYSTEM_RANGE(1, 1) /++ PUBLIC.RANGE_INDEX ++/), (SELECT 2 FROM SYSTEM_RANGE(1, 1) /++ PUBLIC.RANGE_INDEX ++/)) */ WHERE "X" IN((SELECT 1 FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */), (SELECT 2 FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */))
+>> SELECT "SYSTEM_RANGE"."X" FROM SYSTEM_RANGE(1, 10) /* range index: X IN((SELECT 1 FROM SYSTEM_RANGE(1, 1) /++ range index ++/), (SELECT 2 FROM SYSTEM_RANGE(1, 1) /++ range index ++/)) */ WHERE "X" IN((SELECT 1 FROM SYSTEM_RANGE(1, 1) /* range index */), (SELECT 2 FROM SYSTEM_RANGE(1, 1) /* range index */))
