+++ h2/src/test/org/h2/test/scripts/predicates/type.sql
@@ -34,7 +34,7 @@ EXPLAIN SELECT
     'Test' IS OF (VARCHAR), 'Test' IS NOT OF (VARCHAR),
     10 IS OF (VARCHAR), 10 IS NOT OF (VARCHAR),
     NULL IS OF (VARCHAR), NULL IS NOT OF (VARCHAR);
->> SELECT TRUE, FALSE, FALSE, TRUE, UNKNOWN, UNKNOWN FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT TRUE, FALSE, FALSE, TRUE, UNKNOWN, UNKNOWN FROM SYSTEM_RANGE(1, 1) /* range index */
 
 EXPLAIN SELECT A IS OF (INT), A IS OF (BIGINT), A IS NOT OF (INT), NOT A IS OF (BIGINT) FROM TEST;
 >> SELECT ("A" IS OF (INTEGER), ("A" IS OF (BIGINT), ("A" IS NOT OF (INTEGER), ("A" IS NOT OF (BIGINT) FROM "PUBLIC"."TEST" /* PUBLIC.TEST.tableScan */
