+++ h2/src/test/org/h2/test/scripts/predicates/unique.sql
@@ -23,13 +23,13 @@ SELECT UNIQUE(SELECT A, B, C FROM TEST);
 >> TRUE
 
 EXPLAIN SELECT UNIQUE(SELECT A, B FROM TEST);
->> SELECT UNIQUE( SELECT "A", "B" FROM "PUBLIC"."TEST" /* PUBLIC.TEST.tableScan */) FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT UNIQUE( SELECT "A", "B" FROM "PUBLIC"."TEST" /* PUBLIC.TEST.tableScan */) FROM SYSTEM_RANGE(1, 1) /* range index */
 
 SELECT UNIQUE(SELECT A, B FROM TEST);
 >> FALSE
 
 EXPLAIN SELECT UNIQUE(SELECT DISTINCT A, B FROM TEST);
->> SELECT TRUE FROM SYSTEM_RANGE(1, 1) /* PUBLIC.RANGE_INDEX */
+>> SELECT TRUE FROM SYSTEM_RANGE(1, 1) /* range index */
 
 SELECT UNIQUE(SELECT DISTINCT A, B FROM TEST);
 >> TRUE
